Added IncludeShape
